requirements.txt
google-genai
requests
python-dotenv
